added timestamp to report page
