+++ src/dpr/components/async-report-list/view.njk
@@ -1,12 +1,16 @@
 {% from "../async-columns/view.njk" import dprAsyncColumns %}
 {% from "../pagination/view.njk" import dprPagination %}
 {% from "../async-data-table/view.njk" import dprAsyncDataTable %}
+{% from "../async-applied-filters/view.njk" import dprAsyncAppliedFiltersList %}
 {% from "govuk/components/details/macro.njk" import govukDetails %}
 
 {% macro dprAsyncReportList(options) %}
 
   {% set columns = options.columns %}
   {% set pagination = options.pagination %}
+  {% set appliedFilters = options.appliedFilters %}
+
+  {{ dprAsyncAppliedFiltersList(appliedFilters) }}
 
   <div class="columns-select-wrapper">
     {% set columnsHtml %}
